+++ Domain/Repository/ItemRepository.swift
@@ -3,5 +3,5 @@ import Combine
 
 /// @mockable
 public protocol ItemRepository {
-    func fetch() -> AnyPublisher<[Item], Error>
+    func fetch(completion: @escaping (Swift.Result<[Item], Error>) -> Void)
 }
